initial fix logic
